Add plain-auth for svn direct connection
add plain authenticate for compatibility of svn direct connect:  svn uses a plain text passwd file, but if.svnadmin uses an encrypted passwd file , so this patch is to fix this problem. To use this patch, you should use a unique configure file for your svnserve, you can use --config-file option to direct to a unique configure file and share passwd and authz files with if.svnadmin.